Can now simulate extraction

+++ src/toniarts/openkeeper/tools/convert/wad/WadFile.java
@@ -39,7 +39,7 @@ import toniarts.openkeeper.tools.convert.Utils;
  * @author Toni Helenius <helenius.toni@gmail.com>
  */
 public class WadFile {
-    
+
     private final File file;
     private final LinkedHashMap<String, WadFileEntry> wadFileEntries;
     private static final String WAD_HEADER_IDENTIFIER = "DWFB";
@@ -73,7 +73,7 @@ public class WadFile {
 
             //Seek
             rawWad.seek(0x48);
-            
+
             int files = Utils.readUnsignedInteger(rawWad);
             int nameOffset = Utils.readUnsignedInteger(rawWad);
             int nameSize = Utils.readUnsignedInteger(rawWad);
@@ -156,9 +156,9 @@ public class WadFile {
 
         //Open the WAD for extraction
         try (RandomAccessFile rawWad = new RandomAccessFile(file, "r")) {
-            
+
             for (String fileName : wadFileEntries.keySet()) {
-                extractFileData(fileName, destination, rawWad);
+                extractFileData(fileName, destination, rawWad, false);
             }
         } catch (Exception e) {
 
@@ -174,14 +174,14 @@ public class WadFile {
      * @param destination destination directory
      * @param rawWad the opened WAD file
      */
-    private File extractFileData(String fileName, String destination, RandomAccessFile rawWad) {
+    private File extractFileData(String fileName, String destination, RandomAccessFile rawWad, boolean simulation) {
 
         //See that the destination is formatted correctly and create it if it does not exist
         String dest = destination;
         if (!dest.endsWith(File.separator)) {
             dest = dest.concat(File.separator);
         }
-        
+
         String mkdir = dest;
         if (fileName.contains(File.separator)) {
             this.subdir = fileName.substring(0, fileName.lastIndexOf(File.separator) + 1);
@@ -189,10 +189,15 @@ public class WadFile {
         } else {
             dest += this.subdir;
         }
-        
+
         File destinationFolder = new File(mkdir);
         destinationFolder.mkdirs();
         dest = dest.concat(fileName);
+
+        // Simulation?
+        if (simulation) {
+            return new File(dest);
+        }
 
         //Write to the file
         try (OutputStream outputStream = new FileOutputStream(dest)) {
@@ -210,10 +215,22 @@ public class WadFile {
      * @param destination destination directory
      */
     public File extractFileData(String fileName, String destination) {
+        return extractFileData(fileName, destination, false);
+    }
+
+    /**
+     * Extract a single file to a given location
+     *
+     * @param fileName file to extract
+     * @param destination destination directory
+     * @param simulation simulate only, no extraction is done, but set subdir
+     * scheming gets preserved and the directoried get made
+     */
+    public File extractFileData(String fileName, String destination, boolean simulation) {
 
         //Open the WAD for extraction
         try (RandomAccessFile rawWad = new RandomAccessFile(file, "r")) {
-            return extractFileData(fileName, destination, rawWad);
+            return extractFileData(fileName, destination, rawWad, simulation);
         } catch (Exception e) {
 
             //Fug
@@ -236,14 +253,14 @@ public class WadFile {
         if (fileEntry == null) {
             throw new RuntimeException("File " + fileName + " not found from the WAD archive!");
         }
-        
+
         try {
 
             //Seek to the file we want and read it
             rawWad.seek(fileEntry.getOffset());
             byte[] bytes = new byte[fileEntry.getCompressedSize()];
             rawWad.read(bytes);
-            
+
             result = new ByteArrayOutputStream();
 
             //See if the file is compressed
@@ -252,13 +269,13 @@ public class WadFile {
             } else {
                 result.write(bytes);
             }
-            
+
         } catch (Exception e) {
 
             //Fug
             throw new RuntimeException("Failed to read the WAD file!", e);
         }
-        
+
         return result;
     }
 
@@ -319,7 +336,7 @@ public class WadFile {
                 k -= (Utils.toUnsignedByte(flag) & 0x60) << 3;
                 k -= Utils.toUnsignedByte(tmp);
                 k--;
-                
+
                 counter = ((Utils.toUnsignedByte(flag) >> 2) & 7) + 2;
                 do {
                     dest[j] = dest[k++];
